update screen post
