--- src/components/NewPost/styles.ts
@@ -1,19 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-  container: {
-    marginHorizontal: 10,
-    marginTop:10,
-  },
-  headerContainer:{
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  headerText:{
-    color:'black',
-    fontWeight:'700',
-    fontSize: 20,
-    marginRight: 23,
-  },
-});
