+++ src/screens/NewPostScreen/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { SafeAreaView } from 'react-native';
 
-import AddNewPost from '../../components/NewPost/AddNewPost';
+import AddNewPost from './NewPost/AddNewPost';
 import styles from './styles';
 
 const NewPostScreen = () => {
   return (
     <SafeAreaView style={styles.safeArea}>
-        <AddNewPost/>
+      <AddNewPost />
     </SafeAreaView>
   );
 };
